restructure application architecture Relates #25
