+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/martin/.pyenv/versions/billy-search/bin/python"
+}
